refactor out common processing util functions
