+++ modular/phases/actor.py
@@ -11,87 +11,18 @@ sys.path.append(str(Path(__file__).parent.parent.parent))
 from type_defs.base import Node, Option
 from type_defs.operations import (
     BaseOperationRequest,
-    BashParams,
-    BashRequest,
     OperationMetadata,
-    PythonParams,
-    PythonRequest,
-    ScoreLogParams,
-    ScoreLogRequest,
-    ScoreParams,
-    ScoreRequest,
-    SubmissionParams,
-    SubmissionRequest,
 )
 from type_defs.phases import StateRequest
-from type_defs.states import AgentState, ModularState
-from utils.functions import validate_function_call
+from type_defs.states import ModularState
+from utils.functions import (
+    create_standard_tool_operation,
+    handle_set_timeout,
+    validate_function_call,
+)
 from utils.logging import create_log_request, log_warning
 from utils.phase_utils import get_last_completion, run_phase
 from utils.styles import log_styles
-
-
-def create_standard_tool_operation(
-    tool_name: str, tool_args: dict, metadata: OperationMetadata
-) -> BaseOperationRequest | None:
-    if tool_name == "submit":
-        return SubmissionRequest(
-            type="submit",
-            params=SubmissionParams(submission=tool_args["answer"]),
-            metadata=metadata,
-        )
-    elif tool_name == "run_bash":
-        return BashRequest(
-            type="bash",
-            params=BashParams(command=tool_args["command"]),
-            metadata=metadata,
-        )
-    elif tool_name == "run_python":
-        return PythonRequest(
-            type="python",
-            params=PythonParams(code=tool_args["code"]),
-            metadata=metadata,
-        )
-    elif tool_name == "score":
-        return ScoreRequest(type="score", params=ScoreParams(), metadata=metadata)
-    elif tool_name == "score_log":
-        return ScoreLogRequest(
-            type="score_log", params=ScoreLogParams(), metadata=metadata
-        )
-    else:
-        return None
-
-
-def handle_set_timeout(state: AgentState, tool_args: dict) -> AgentState:
-    """Handle the set_timeout tool operation"""
-    try:
-        state.timeout = int(tool_args["timeout"])
-        state.nodes.append(
-            Node(
-                source="tool_output",
-                options=[
-                    Option(
-                        content=f"Timeout set to {state.timeout}", name="set_timeout"
-                    )
-                ],
-            )
-        )
-    except (KeyError, ValueError):
-        state.nodes.append(
-            Node(
-                source="warning",
-                options=[
-                    Option(
-                        content=(
-                            "Invalid set_timeout function call, timeout remains ",
-                            f"{state.timeout} seconds",
-                        )
-                    )
-                ],
-            )
-        )
-    state.update_usage()
-    return state
 
 
 def create_function_call_log_message(
